fix: исправлена мидлвара cors
